requirements.txt
anyio==4.7.0
certifi==2024.12.14
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
python-telegram-bot==20.3
sniffio==1.3.1
typing_extensions==4.12.2
